done: expense

+++ branches/models.py
@@ -25,7 +25,7 @@ class Branch(models.Model):
     name = models.CharField(max_length=255, verbose_name=_('Name'))
     address = models.TextField(blank=True, null=True, verbose_name=_('Address'))
     phone_number = models.CharField(max_length=13, blank=True, null=True, verbose_name=_('Phone number'))
-    balance = models.FloatField(verbose_name=_('Balance'), default=0)
+    balance = models.DecimalField(max_digits=15, decimal_places=2, verbose_name=_('Balance'), default=0)
     created_at = models.DateTimeField(auto_now_add=True, verbose_name=_('Created at'))
 
     class Meta:
